requirements.txt
pandas
python-dotenv
SQLAlchemy
PyMySQL
torch
scikit-learn
matplotlib